==> notebooks/0.0-test-environtment.ipynb <==
import sys
sys.version

%pwd

import numpy
import pandas 
import matplotlib.pyplot as plt
import matplotlib
import sklearn

print(numpy.__version__)
print(pandas.__version__)
print(matplotlib.__version__)
print(numpy.__version__)
