==> notebooks/4.0-modeling-classification.ipynb <==
import warnings
import numpy as np
import pandas as pd
from sklearn.impute import SimpleImputer
from sklearn.model_selection import StratifiedShuffleSplit
from sklearn.pipeline import Pipeline
from sklearn.preprocessing import StandardScaler
from sklearn.preprocessing import OrdinalEncoder, OneHotEncoder, LabelEncoder
from sklearn.compose import ColumnTransformer
from sklearn.model_selection import train_test_split

#supressing warnings for readability
warnings.filterwarnings("ignore")

# To plot pretty figures directly within Jupyter
%matplotlib inline

# pandas options
pd.set_option("display.max.columns", None)
pd.set_option("display.max.rows", None)
pd.set_option("display.precision", 2)

import pickle

with open('../data/var_store/X_resampled.pkl', 'rb') as f:
    X_resampled = pickle.load(f)

with open('../data/var_store/y_resampled.pkl', 'rb') as f:
    y_resampled = pickle.load(f)

from sklearn.tree import DecisionTreeClassifier
from sklearn.metrics import classification_report

X_train, X_test, y_train, y_test = train_test_split(X_resampled, y_resampled, test_size=0.2, random_state=42)

# Create a decision tree classifier
tree = DecisionTreeClassifier()

# Train the decision tree classifier with class weighting
tree.fit(X_train, y_train)

# Make predictions on the test data
predictions = tree.predict(X_test)

# Evaluate the model performance
print(classification_report(y_test, predictions))

from sklearn.metrics import accuracy_score
print(accuracy_score(y_test, predictions))

import optuna
import gc
import lightgbm as lgb
from optuna import Trial
from sklearn.model_selection import KFold
from lightgbm import LGBMClassifier
from sklearn.metrics import roc_auc_score

def fitLGBM(trial,X, y):
    
    params={
      'n_estimators':trial.suggest_int('n_estimators', 0, 1000), 
      'num_leaves':trial.suggest_int('num_leaves', 2, 512),
      'max_depth':trial.suggest_int('max_depth', 2, 128),
      'learning_rate': trial.suggest_loguniform('learning_rate', 0.001, 0.15),
      'min_split_gain': trial.suggest_loguniform('min_split_gain', 0.001, 0.1),
      'feature_fraction':trial.suggest_uniform('feature_fraction',0.1, 1.0),
      'bagging_freq':trial.suggest_int('bagging_freq',0.1,10),
      'verbosity': -1,
      'random_state': 42
            }
    kfolds = KFold(n_splits=5, shuffle=True, random_state=42)
    model = LGBMClassifier(**params)
    
    res=[]
    for i, (tdx, vdx) in enumerate(kfolds.split(X, y)):
        X_train, X_valid, y_train, y_valid = X[tdx], X[vdx], y[tdx], y[vdx]
        model.fit(X_train, y_train,
                 eval_set=[(X_train, y_train), (X_valid, y_valid)],
                 early_stopping_rounds=30, verbose=False)
        preds = model.predict_proba(X_valid)
        res.append(roc_auc_score(y_valid, preds[:,1]))
    
    err = np.mean(res)
    
    return model, err

def objective(trial:Trial):
    
    gc.collect()
    models=[]
    validScore=0
   
    model,log = fitLGBM(trial,X_resampled,y_resampled)
    
    models.append(model)
    gc.collect()
    validScore+=log
    validScore/=len(models)
    
    return validScore

study = optuna.create_study(direction='maximize', pruner=optuna.pruners.MedianPruner(n_warmup_steps=5))
study.optimize(objective, timeout=60*60*2)

best_param = study.best_params
print(study.best_value)
best_param

model = LGBMClassifier(**best_param)

%%time
model.fit(X_train, y_train)

y_pred_val = model.predict(X_test)
accuracy_val = accuracy_score(y_test, y_pred_val)
print(f"Validation Accuracy: {accuracy_val}")

# model.booster_.save_model('../models/1-cls-07032023.txt')

loaded_model = lgb.Booster(model_file='../models/1-cls-07032023.txt')

y_pred_val = loaded_model.predict(X_test)
accuracy_val = accuracy_score(y_test, y_pred_val > 0.5)
print(f"Validation Accuracy: {accuracy_val}")
